+++ src/gl9_scene/camera.cpp
@@ -13,7 +13,7 @@ void Camera::update() {
   viewMatrix = lookAt(position, position-back, up);
 }
 
-glm::vec3 Camera::cast(float u, float v) {
+glm::vec3 Camera::cast(double u, double v) {
   // Create point in Screen coordinates
   glm::vec4 screenPosition{u,v,0.0f,1.0f};
